Fix image updating issue

See merge request acac/acac-v14!12

+++ local-addon/pm_culinary/models/pm_ingredient.py
@@ -87,7 +87,7 @@ class ProductTemplate(models.Model):
 
     def write(self, vals):
       # Temporarily fixing image issue when update a record
-      if vals['image_1920']:
+      if 'image_1920' in vals and vals['image_1920']:
           # print(self._name)
           # print(self.id)
           self.env.cr.execute("""DELETE FROM ir_attachment WHERE res_model = '%s' AND res_id = %d""" % (self._name, self.id))
