+++ local-addon/pm_culinary/models/pm_recipe.py
@@ -134,7 +134,7 @@ class PmRecipe(models.Model):
         vals['is_expired'] = False
 
         # Temporarily fixing image issue when update a record
-        if vals['recipe_image']:
+        if 'recipe_image' in vals and vals['recipe_image']:
             self.env.cr.execute("""DELETE FROM ir_attachment WHERE res_model = '%s' AND res_field = '%s' AND res_id = %d""" % (self._name, 'recipe_image', self.id))
 
         return super().write(vals)
